Refactor tool installation code in autoOS
