+++ agent-package-manager/src/utils/commands/install/toolInstall.ts
@@ -1,13 +1,12 @@
 import fs from 'fs/promises';
 import path from 'path';
-import extract from 'extract-zip';
-import ora from 'ora';
-import { getToolInstallDir, PACKAGES_DIR } from '../../shared/path.js';
-import { getToolFromRegistry, getToolVersionFromRegistry } from '../registry/toolInquiry.js';
+import { PACKAGES_DIR } from '../../shared/path.js';
 import { downloadFileFromDsn } from '../../autoDrive/autoDriveClient.js';
-import { InstallOptions, ToolInstallInfo, ToolMetadata } from '../../../types/index.js';
+import { ToolInstallInfo } from '../../../types/index.js';
 import { loadCredentials } from '../../credential/index.js';
 import { getCidFromHash } from '../../blockchain/utils.js';
+import extract from "extract-zip";
+
 
 const fetchToolPackage = async (cidHash: string): Promise<string> => {
   const cid = getCidFromHash(cidHash);
@@ -51,18 +50,15 @@ const unpackToolToDirectory = async (
     console.error('Error extracting package:', error);
     throw error;
   }
-};
+};    
 
 const performToolInstallation = async (
   toolInfo: ToolInstallInfo,
+  toolInstallDir: string,
 ): Promise<string> => {
   try {
-    const { installDir } = await getToolInstallDir();
-    if (!installDir) {
-      throw new Error('Could not detect project root. Make sure you are in a project directory.');
-    }
     const packagePath = await fetchToolPackage(toolInfo.cid);
-    const toolDir = await unpackToolToDirectory(packagePath, toolInfo.name, installDir);
+    const toolDir = await unpackToolToDirectory(packagePath, toolInfo.name, toolInstallDir);
     console.log(`Tool installed successfully to: ${toolDir}`);
     return toolDir;
   } catch (error) {
@@ -71,127 +67,8 @@ const performToolInstallation = async (
   }
 };
 
-const validateCidOption = (cid: string | undefined): string => {
-  if (!cid) {
-    throw new Error('CID is required when installing a tool by CID');
-  }
-  return cid;
-};
-
-const validateVersionOption = (version: string | undefined): string => {
-  if (!version) {
-    throw new Error('Version is required when installing a specific version');
-  }
-  return version;
-};
-
-export const createToolInfoFromRegistry = (registryInfo: ToolMetadata): ToolInstallInfo => {
-  return {
-    name: registryInfo.name,
-    cid: registryInfo.cid,
-    version: registryInfo.version,
-  };
-};
-
-const resolveCidInstallation = (
-  toolName: string,
-  cid: string,
-): {
-  toolInfo: ToolInstallInfo;
-  versionDisplay: string;
-} => {
-  return {
-    toolInfo: {
-      name: toolName,
-      cid,
-    },
-    versionDisplay: '',
-  };
-};
-
-const resolveVersionInstallation = async (
-  toolName: string,
-  version: string,
-  spinner: ReturnType<typeof ora>,
-): Promise<{ toolInfo: ToolInstallInfo; versionDisplay: string }> => {
-  const versionDisplay = `version ${version}`;
-  spinner.text = `Looking for ${toolName} ${versionDisplay}...`;
-
-  const registryToolInfo = await getToolVersionFromRegistry(toolName, version);
-  if (!registryToolInfo) {
-    throw new Error(
-      `${versionDisplay} of tool '${toolName}' not found in registry. Use 'autoOS list -d' to see available versions.`,
-    );
-  }
-
-  return {
-    toolInfo: createToolInfoFromRegistry(registryToolInfo),
-    versionDisplay,
-  };
-};
-
-const resolveLatestInstallation = async (
-  toolName: string,
-  spinner: ReturnType<typeof ora>,
-): Promise<{ toolInfo: ToolInstallInfo; versionDisplay: string }> => {
-  spinner.text = `Looking for latest version of ${toolName}...`;
-  const registryToolInfo = await getToolFromRegistry(toolName);
-
-  if (!registryToolInfo) {
-    throw new Error(`Tool '${toolName}' not found in registry`);
-  }
-
-  const versionDisplay = `(latest: ${registryToolInfo.version})`;
-
-  return {
-    toolInfo: createToolInfoFromRegistry(registryToolInfo),
-    versionDisplay,
-  };
-};
-
-const updateSpinnerWithInstallInfo = (
-  spinner: ReturnType<typeof ora>,
-  toolName: string,
-  versionDisplay: string,
-  installType: string,
-): void => {
-  spinner.text = `Installing ${toolName} ${versionDisplay} ${installType} from registry...`;
-};
-
-const resolveToolInfo = async (
-  toolName: string,
-  options: InstallOptions,
-  spinner: ReturnType<typeof ora>,
-): Promise<{ toolInfo: ToolInstallInfo; versionDisplay: string }> => {
-  // All installations are local now
-  const installType = 'locally';
-
-  // Handle CID-based installation
-  if (options.cid) {
-    const validCid = validateCidOption(options.cid);
-    spinner.text = `Installing ${toolName} ${installType} using CID: ${validCid}`;
-    return resolveCidInstallation(toolName, validCid);
-  }
-
-  // Handle version-based installation
-  if (options.version) {
-    const validVersion = validateVersionOption(options.version);
-    const result = await resolveVersionInstallation(toolName, validVersion, spinner);
-    updateSpinnerWithInstallInfo(spinner, toolName, result.versionDisplay, installType);
-    return result;
-  }
-
-  // Handle latest version installation (default)
-  const result = await resolveLatestInstallation(toolName, spinner);
-  updateSpinnerWithInstallInfo(spinner, toolName, result.versionDisplay, installType);
-  return result;
-};
-
 export {
   fetchToolPackage,
   unpackToolToDirectory,
   performToolInstallation,
-  validateCidOption,
-  validateVersionOption,
-  resolveToolInfo,
 };
